feat: 로그인 연장 테스트 버튼을 대쉬보드 페이지로 이동
